Rotate gallery images on upload
JPEG pictures can be rotated and have an EXIF header describing which
way to rotate them which most image viewers respect.

Unfortunately, browsers are not such image viewers, so rotate the
image correctly according to that tag and remove it after rotation.

This means requiring the piexif library which can be used for reading
the timestamp data as well, so we don't need the exifread library
anymore.