Added booking system
Make use of the Booking model by creating a form for creating a
booking when a date is clicked in the calendar view.

Make sure the booking request is somewhat sane by ensuring that it
doesn't start before it ends and that it is not already booked and add
tests for that.