Minor improvements to photo gallery
Give the date and description of images a black shadow so they are
easier to read on a light image.

Move javascript code to the top of the HTML file where it belongs.

Use HTML escape codes for image navigation instead of UTF-8 for easier
readability.

Order images by date in gallery.