Use bootstrap datepicker
Instead of using jquery-ui for displaying a datepicker, use bootstrap
datepicker to get a more consistent look as well as lowering
dependencies.

This can also be used to select a range of dates, which is hopefully
more logical since a range of dates is exactly what needs to be
selected.