Ensure image file upload in gallery admin is multipart
Since the form for uploading multiple images in the gallery from the
Django admin form is customized, Django cannot detect that it needs to
be a multipart form causing the image uploads to fail silently.

As a bit of a hack, ensure that the form returns that it is indeed a
multipart form to ensure it is rendered correctly.