Make it possible to edit existing booking
If the booking selected is either owned by the current user or the
current user is admin, instead of simply displaying the booking make
it editable using the same form as if it was a new booking.

This also means redirecting to the calendar page after
creating/updating a booking as anything else would be rather
confusing.